requirements.txt
torch --index-url https://download.pytorch.org/whl/cu121
transformers>=4.37.0
accelerate>=0.26.0
langchain>=0.3.0
langchain-community>=0.2.0
sentencepiece
protobuf==4.25.3
python-dotenv

langchain-huggingface
transformers
huggingface-hub

langchain-core

streamlit==1.49.1